+++ src/days/day3.rs
@@ -1,17 +1,17 @@
 use regex::Regex;
 
-pub fn part1(input: &str) -> Result<i32, &'static str> {
+pub fn part1(input: &str) -> Result<isize, &'static str> {
     let mut res = 0;
     let re = Regex::new(r"mul\((\d{1,3}),(\d{1,3})\)").unwrap();
     for cap in re.captures_iter(input) {
-        let a = cap[1].parse::<i32>().unwrap();
-        let b = cap[2].parse::<i32>().unwrap();
+        let a = cap[1].parse::<isize>().unwrap();
+        let b = cap[2].parse::<isize>().unwrap();
         res += a * b;
     }
     Ok(res)
 }
 
-pub fn part2(input: &str) -> Result<i32, &'static str> {
+pub fn part2(input: &str) -> Result<isize, &'static str> {
     let re = Regex::new(r"(?s:don't\(\).*?(do\(\)|$))").unwrap();
     let parsed = re.replace_all(input, "");
     part1(&parsed)
